Refactor: extract common code logic for helm-type plugins
