+++ internal/pkg/plugin/jenkins/read.go
@@ -6,6 +6,7 @@ import (
 	"github.com/mitchellh/mapstructure"
 	"k8s.io/utils/strings/slices"
 
+	. "github.com/merico-dev/stream/internal/pkg/plugin/common/helm"
 	"github.com/merico-dev/stream/pkg/util/helm"
 	"github.com/merico-dev/stream/pkg/util/k8s"
 	"github.com/merico-dev/stream/pkg/util/log"
