+++ internal/pkg/plugin/jenkins/validation.go
@@ -1,6 +1,9 @@
 package jenkins
 
-import "github.com/merico-dev/stream/pkg/util/helm"
+import (
+	. "github.com/merico-dev/stream/internal/pkg/plugin/common/helm"
+	"github.com/merico-dev/stream/pkg/util/helm"
+)
 
 // validate validates the options provided by the core.
 func validate(param *Param) []error {
