+++ internal/pkg/plugin/kubeprometheus/delete.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/mitchellh/mapstructure"
 
+	. "github.com/merico-dev/stream/internal/pkg/plugin/common/helm"
 	"github.com/merico-dev/stream/pkg/util/helm"
 	"github.com/merico-dev/stream/pkg/util/k8s"
 	"github.com/merico-dev/stream/pkg/util/log"
